+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 ### Features
 
  * Add tabs plugin to support syntax show content
+ * Add vite-plugin-svg plugin documentation